requirements.txt
platformdirs>=2.2.0
PyYAML==6.0.1
playwright==1.37.0
ultralytics==8.0.161
Pillow==9.3.0
paddleocr==2.7.0.2
numpy==1.25.2
easyocr
sentencepiece
transformers==4.30.2
protobuf
accelerate
cpm_kernels
torch>=2.0
gradio